+++ src/main/java/com/metacrew/pr2s/repository/joininforepository/JoinInfoQueryRepository.java
@@ -1,4 +1,4 @@
-package com.metacrew.pr2s.repository.joinforepository;
+package com.metacrew.pr2s.repository.joininforepository;
 
 import com.metacrew.pr2s.dto.condition.JoinInfoConditionDto;
 import com.metacrew.pr2s.entity.JoinInfo;
