+++ src/test/java/com/metacrew/pr2s/repository/joininforepository/JoinInfoRepositoryTest.java
@@ -7,7 +7,6 @@ import com.metacrew.pr2s.dto.WorkdaysDto;
 import com.metacrew.pr2s.entity.*;
 import com.metacrew.pr2s.repository.AddressRepository;
 import com.metacrew.pr2s.repository.institutionrepository.InstitutionRepository;
-import com.metacrew.pr2s.repository.joinforepository.JoinInfoRepository;
 import com.metacrew.pr2s.repository.memberrepository.MemberRepository;
 import com.metacrew.pr2s.repository.workdaysrepository.WorkdaysRepository;
 import org.junit.jupiter.api.Test;
