+++ src/test/java/com/metacrew/pr2s/service/memberservice/ClientMemberServiceTest.java
@@ -4,7 +4,7 @@ import com.metacrew.pr2s.dto.*;
 import com.metacrew.pr2s.entity.*;
 import com.metacrew.pr2s.repository.FileRepository;
 import com.metacrew.pr2s.repository.institutionrepository.InstitutionRepository;
-import com.metacrew.pr2s.repository.joinforepository.JoinInfoRepository;
+import com.metacrew.pr2s.repository.joininforepository.JoinInfoRepository;
 import com.metacrew.pr2s.repository.memberrepository.MemberRepository;
 import com.metacrew.pr2s.repository.workdaysrepository.WorkdaysRepository;
 import com.metacrew.pr2s.service.institutionservice.InstitutionService;
